% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{K-means Cluster with Coordinate Descent (E-M)}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{A matrix with samples stored in rows (dim: \code{n * p}).}

\item{K}{An integer representing the number of cluster.}

\item{M}{A matrix with initial centroids stored in rows (dim: \code{K * p}, optional).}

\item{numIter}{The number of iterations.}
}
\value{
A vector with integers from \code{1} to \code{K}, representing the cluster each sample belongs to.
}
\description{
K-means Cluster with Coordinate Descent (E-M)
}
\examples{
# Kmeans Algorithm Implementation Example.1
set.seed(123)
X1 <- matrix(rnorm(100), nrow = 20)
MyKmeans(X1, K = 3)

# Kmeans Algorithm Implementation Example.2
set.seed(456)
X2 <- matrix(rnorm(100), nrow = 20)
MyKmeans(X2, K = 3)
}
