% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{K-means Cluster with Coordinate Descent}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{A matrix with samples stored in rows (dim: \code{n * p}).}

\item{K}{An integer representing the number of cluster.}

\item{M}{A matrix with initial centroids stored in rows (dim: \code{K * p}, optional).}

\item{numIter}{The number of iterations.}
}
\value{
A vector with integers from \code{1} to \code{K}, representing the cluster each sample belongs to.
}
\description{
A function that performs K-means clustering to partition n data points into K clusters.
For each iteration, it assigns each data point to its nearest centroid based on 
Euclidean distance, then updates the cluster centroids by calculating the mean of 
all points in each cluster. The process continues until either convergence is reached
or the maximum number of iterations is met. The core computation is implemented in 
C++ for efficiency.
}
\examples{
# Kmeans Algorithm Implementation Example.1
set.seed(123)
X1 <- matrix(rnorm(100), nrow = 20)
MyKmeans(X = X1, K = 3)

}
